pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lattice"
version = "0.1.0"
description = "Quant pricing and trading library built on dag and livetable"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Abhishek Gulati"}
]
keywords = [
    "quant",
    "finance",
    "trading",
    "pricing",
    "options",
    "derivatives",
    "dag",
    "reactive",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "dag-framework>=0.1.0",
    "livetable>=0.1.0",
    "numpy>=1.20",
]

[project.urls]
Homepage = "https://github.com/beeblebroxz/lattice"
Repository = "https://github.com/beeblebroxz/lattice"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
